Upgraded to Flask
Strategic Goal:
Need to receive the token by creating an HTTPS endpoint for token provider to hit.
Also need to save the token outside the context of the request to be used by the main module.

Tactical Implementation:
Python's built-in http module creates a new instance of a handler for each REST call which means I had to save the token to file system and read it back.
This was cumbersome, so I switched to Flask which still allows easy SSL and allows me to pass the self context to the handler.
The only complicated part is starting the server in a different thread, but blocking the main thread until the server receives the token.
Afterwards, a REST call is sent to shutdown the server while the main program continues. This improves speed because the main thread
doesn't have to wait for the server to shutdown.

Other Notes:
    -  added basic dict caching with default caching of 60 seconds. This is easily modifiable by setting the appropriate session headers.
    -  fixed token being overwritten if it already exists using 'x' flag in open. Only supported in Python 3.3 or greater. I'm okay with that.